More _mm_mul_epu32 based changes.
Implement and default to a C version.
Leave the skeleton of the neon version in place, its unclear if it can be made to work but if it can at least it serves as a starting point for someone else.
Implement more comprehensive test hardness for _mm_mul_epu32, this will assist with any future attempts at a neon implementation.