Add more shifts, workaround Clang pickiness
The _mm_sll and _mm_sllv families are now implemented.

Additionally, I switched to using a variable vshlq on the immediate
values, since Clang is pickier about what is a constant and what isn't.
This makes it more flexible, and easier - NEON and SSE behave similarly
with >bitsof(T) shifts.