Change __m128i from int32x4_t to int64x2_t
GCC and Clang allow SIMD vectors to be operated on like integers and
indexed like arrays, similar to Altivec's behavior.

GCC and Clang both internally define __m128i as a vector of 2 long
longs, and when used with the extension, they act this way.

Therefore, to match this behavior, __m128i should be defined as an
int64x2_t.

This also fixes some missing/incorrect vreinterprets.