Reimplement accurate _mm_sqrt_ps for A64
Aarch64 has an accurate instruction vsqrtq_f32, which calculates
the square root for each vector element in the source SIMD and FP
register, places the result in a vector, and writes the vector to
the destination SIMD and FP register. Intrinsic _mm_sqrt_ps can be
implemented with this instruction for better accuracy.