Drop incorrect _mm_srai_epi32 implementation
There were two _mm_srai_epi32 implementations: one was the inline
function; another was the macro. According to Intel manual, the
operation of intrinsic _mm_srai_epi32 is illustrated as following:

__m128i _mm_srai_epi32 (__m128i a, int imm8)
FOR j := 0 to 3
	i := j*32
	IF imm8[7:0] > 31
		dst[i+31:i] := (a[i+31] ? 0xFFFFFFFF : 0x0)
	ELSE
		dst[i+31:i] := SignExtend32(a[i+31:i] >> imm8[7:0])
	FI
ENDFOR

The original inlined function would produce different behavior, and
this patch drops it for the sake of correctness.