Simplify test suite generation process
Closes #72.
The current test suite generation is too complicated, we need to change
three separated places to add a single test.

This commit attempted to simply the process when generating test suites.
The changes in short are as follows:
1. Add option to indicate if a test case is still not implemented.
2. Integrate `InstructionTest`, `getInstructionString`, and
   `runSingleTest` into macros for better management.

Since the test cases are indicated unimplemented only noticed as comments in
the source files, it is hard to track whether an intrinsic's test case
is implemented or not. Therefore, I modified the return value from
`bool` to `result_t` to accept three states: TEST_SUCESS(1),
TEST_FAIL(0), and TEST_UNIMPL(-1). In this way `main.cpp` can recognize
whether this test case is implemented or not and print out the result.

This commit also made some changes to have the tests look cleaner, which
might be easier for contributors to add new test cases.

First, Some common functions in `impl` files are moved to `common` files.
Second, `getInstructionString` and `runSingleTest` was previously implemented
using switch statements, which makes the code lengthy. These two
components and `enum InstructionTest` look similar, so I think it would
be better to implement them as macros.

The proposed method to combine `InstructionTest`, `runSingleTest`, and
`getInstructionString` is mentioned in this issue comment:
https://github.com/DLTcollab/sse2neon/issues/72#issuecomment-672015330