refactor: Wrap _mm_load_pd for unification
The wrapping is for the functions which do the similar things.
For example, do_mm_load_ps() is for loading the __m128 or __m128i from the address.
The new added do_mm_load_pd() is for loading the __m128d data from the address.
In the future, loading the __m64 data from the address would be added,
which does not have the direct mapping intrinsic.