fix: Add correct GCC version checking of vld1q_u8_x4
The comment of GCC bug in 2021-01-12 said that the intrinsic
vld1q_u8_x4 is not implemented in GCC.
Ref:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=71233#c73

However, the GCC 10.3 released in 2021-04-08 has fixed the problem for
AArch64 CPU architecture, not for ARM CPU architecture.
Ref:
https://github.com/gcc-mirror/gcc/commit/391625888d4d97f9016ab9ac04acc55d81f0c26f

The Github Actions CI has change the GCC compiler from 10.1 to 10.3,
hence we should take care of the version well.