test: Change loading functions to non-aligned version
The functions for loading the __m64, __m128, __128i and __m128d data
from address are changed to the version which does not need to be
aligned on any particular boundary.
The name of the functions are changed to a more intuitive naming and do not
bound with the intrinsic name.