feat: Implement _MM_[GET|SET]_DENORMALS_ZERO_MODE
X86 has two modes, which are flush to zero mode and denormals are zeros
mode.
ARM has one mode only, which is flush to zero mode.
The ARM flush to zero mode is like the combination of the two modes of
X86.

However, AArch32 Advanced SIMD arithmetic always uses the flush to zero
setting, regardless of the value of the FZ(flush to zero) bit.

The implementation would change the FZ bit and the testing would focus
on whether the input denormal floating point is treated as zero or not.

Close #465.