refactor: Rewrite _mm_alignr_epi8() to function
The test has been rewritten for testing more immediate values.
The selected values are able to represent different scenarios.

The implementation replaces vextq_u8() with vld1q_u8() since the latter
does not require constant immediate value.
As the immediate value is not required, the implementation can be
changed from macro to function.

The commit fixes the issue #482 as well.

Close #482.