fix: Fix the intrinsics of comparing floating points
Take not greater than (ngt) as an example:
When comparing with NaN values, not greater than (ngt) is not equal
to less equal (le).
The implementation and testing of not greater than (ngt) should be the reverse
of greater than (gt).
Hence, the testing should add the input data of NaN values and the
implementation and testing should be rewritten.