Use `frint32x` (round-to-integral) for _mm_cvtp[s,d]_epi32 if supported
Use the frint32x instruction (FEAT_FRINTTS) for _mm_cvtps_epi32 and
_mm_cvtpd_epi32. This respects the currently set rounding mode.

The round-to-integral instructions match the behaviour specified by the
corresponding SSE instructions; floats that would not fit into 32-bits
instead return the indefinite integer (INT32_MIN). The current
implementation uses a C-style cast, which will saturate out-of-bound
results instead of returning the indefinite integer.

The test for _mm_cvtpd_epi32 has therefore been modified to check for
the integer indefinite for out-of-bound inputs in the case that these
instructions are supported on the -march target.

Example codegen for _mm_cvtpd_epi32(a) with GCC 11.2.0 (-O3,
-march=armv8.5-a):

Prior to commit:
    frinti   v0.2d, v0.2d
    sub      sp, sp, 0x10
    str      xzr, [sp, 8]
    fmov     d1, d0
    mov      d0, v0.d[1]
    fcvtzs   w1, d1
    fcvtzs   w0, d0
    stp      w1, w0, [sp]
    ldr      q0, [sp]
    add      sp, sp, 0x10

After commit:
    frint32x v0.2d, v0.2d
    fcvtzs   v0.2d, v0.2d
    xtn      v0.2s, v0.2d
    mov      d0, v0.d[0]