Add A32 support in CI (#620)
CI: Consider A32 mode in Armv8-A

This commit takes the A32 mode of Armv8-A into consideration. The Arm
Cortex-A32 is chosen as the test target in the CI pipeline because it
only supports AArch32. Moreover, the Cortex-A32 features the Armv8-A
Cryptographic Extension, making it an ideal target for testing Armv8
Cryptographic Extension conversions.