Fix test_mm_dp_pd failure when compiling with -O2 (#643)
The test_mm_dp_pd failed when running "make check" with -O2 passed to
ARCH_CFLAGS. Debug investigation revealed that a buffer initialized with
floats was being  cast to a buffer of doubles. The test failure occurred
due to differences  between the dot product obtained with intrinsics and
the one calculated with C++ doubles. This discrepancy arose because the
numbers were too large, causing rounding errors in the dot product
calculation.

Co-authored-by: Alexander Orlov <aorlov@aligntech.com>