feat: Allow optimization option in sse2neon.h
Revert the previous restriction that some of the functions are forced
to not be optimized when compiling-time optimization options were
given.

Now it is users' responsibility to ensure the behavior after
optimization.

Shifting the responsibility to the users enables sse2neon the run in
the optimized state in general, but not restricted by some specific
scenarios.