Optimize CRC-32C base macro
Optimize CRC-32C base macro when using Armv8-A 32-bit platform.

In brief, we shift left the XOR'ed input message by the size
of each element of the input message.
Hence, after doing carry-less multiplication,
the intermetidate results of CRC always appear in the upper half
of vector, eliminating the needs of masking and subtraction.